Moved FileDetails components into separate controls.
