+++ MediaFinder/Views/Export/ImageViewerView.xaml
@@ -1,0 +1,31 @@
+﻿<UserControl x:Class="MediaFinder.Views.Export.ImageViewerView"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:converters="clr-namespace:MediaFinder.Controls.Wpf.Converters;assembly=MediaFinder.Controls.Wpf"
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+             xmlns:models="clr-namespace:MediaFinder.Models;assembly=MediaFinder.Models"
+             d:DataContext="{d:DesignInstance Type=models:MediaFile,
+                                              IsDesignTimeCreatable=False}"
+             d:DesignHeight="800"
+             d:DesignWidth="450"
+             mc:Ignorable="d">
+  <UserControl.Resources>
+    <converters:PathToBitmapImageConverter x:Key="PathToBitmapImageConverter" DecodePixelWidth="400" />
+  </UserControl.Resources>
+  <Border Margin="0,0,0,8"
+          Padding="8"
+          HorizontalAlignment="Stretch"
+          Background="{DynamicResource PrimaryHueMidForegroundBrush}"
+          BorderBrush="{DynamicResource PrimaryHueMidBrush}"
+          BorderThickness="1"
+          ClipToBounds="True"
+          CornerRadius="8">
+    <StackPanel>
+      <Image MaxWidth="400"
+             MaxHeight="700"
+             Source="{Binding FilePath, Converter={StaticResource PathToBitmapImageConverter}}"
+             Stretch="Uniform" />
+    </StackPanel>
+  </Border>
+</UserControl>
