+++ MediaFinder/Views/Export/MediaPlayerView.xaml
@@ -1,0 +1,101 @@
+﻿<UserControl x:Class="MediaFinder.Views.Export.MediaPlayerView"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+             xmlns:materialDesign="http://materialdesigninxaml.net/winfx/xaml/themes"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+             xmlns:models="clr-namespace:MediaFinder.Models;assembly=MediaFinder.Models"
+             xmlns:system="clr-namespace:System;assembly=mscorlib"
+             d:DataContext="{d:DesignInstance Type=models:MediaFile,
+                                              IsDesignTimeCreatable=False}"
+             d:DesignHeight="450"
+             d:DesignWidth="800"
+             mc:Ignorable="d">
+  <UserControl.Resources>
+    <system:Double x:Key="IconSize">24</system:Double>
+  </UserControl.Resources>
+  <Border Margin="0,0,0,8"
+          Padding="8"
+          HorizontalAlignment="Stretch"
+          Background="{DynamicResource AttentionToActionBrush}"
+          BorderBrush="{DynamicResource PrimaryHueMidBrush}"
+          BorderThickness="1"
+          ClipToBounds="True"
+          CornerRadius="8">
+    <StackPanel>
+      <MediaElement x:Name="fileVideoViewer"
+                    MaxWidth="400"
+                    Loaded="FileVideoViewer_Loaded"
+                    LoadedBehavior="Manual"
+                    MediaEnded="FileVideoViewer_MediaEnded"
+                    ScrubbingEnabled="True"
+                    Source="{Binding FilePath}"
+                    Stretch="Uniform"
+                    UnloadedBehavior="Stop" />
+
+      <StackPanel Margin="0,8,0,0"
+                  HorizontalAlignment="Center"
+                  Orientation="Horizontal">
+        <!-- Play button. -->
+        <Button x:Name="mediaPlayButton"
+                Margin="5"
+                HorizontalAlignment="Center"
+                Click="PlayButton_Click"
+                Style="{StaticResource MaterialDesignFloatingActionMiniButton}"
+                ToolTip="Play">
+          <materialDesign:PackIcon Width="{StaticResource IconSize}"
+                                   Height="{StaticResource IconSize}"
+                                   HorizontalAlignment="Center"
+                                   VerticalAlignment="Center"
+                                   Kind="PlayCircle" />
+        </Button>
+
+        <!-- Pause button. -->
+        <Button x:Name="mediaPauseButton"
+                Margin="5"
+                HorizontalAlignment="Center"
+                Click="PauseButton_Click"
+                IsEnabled="{Binding IsEnabled, ElementName=mediaPlayButton, Converter={StaticResource InvertBooleanConverter}}"
+                Style="{StaticResource MaterialDesignFloatingActionMiniButton}"
+                ToolTip="Pause">
+          <materialDesign:PackIcon Width="{StaticResource IconSize}"
+                                   Height="{StaticResource IconSize}"
+                                   HorizontalAlignment="Center"
+                                   VerticalAlignment="Center"
+                                   Kind="PauseCircle" />
+        </Button>
+
+        <!-- Stop button. -->
+        <Button Margin="5"
+                HorizontalAlignment="Center"
+                Click="StopButton_Click"
+                IsEnabled="{Binding IsEnabled, ElementName=mediaPlayButton, Converter={StaticResource InvertBooleanConverter}}"
+                Style="{StaticResource MaterialDesignFloatingActionMiniButton}"
+                ToolTip="Stop">
+          <materialDesign:PackIcon Width="{StaticResource IconSize}"
+                                   Height="{StaticResource IconSize}"
+                                   HorizontalAlignment="Center"
+                                   VerticalAlignment="Center"
+                                   Kind="StopCircle" />
+        </Button>
+
+        <!-- speed slider. This slider allows you to change the speed of the media playback. -->
+        <materialDesign:PackIcon Width="{StaticResource IconSize}"
+                                 Height="{StaticResource IconSize}"
+                                 Margin="5"
+                                 VerticalAlignment="Center"
+                                 Kind="PlaySpeed" />
+        <Slider Name="speedRatioSlider"
+                MinWidth="180"
+                HorizontalAlignment="Stretch"
+                VerticalAlignment="Center"
+                IsSnapToTickEnabled="True"
+                Maximum="4"
+                Minimum="0.5"
+                TickFrequency="0.5"
+                TickPlacement="BottomRight"
+                Value="{Binding SpeedRatio, ElementName=fileVideoViewer, Mode=TwoWay}" />
+      </StackPanel>
+    </StackPanel>
+  </Border>
+</UserControl>
